+++ plugins/wikitext/src/01-doc.js
@@ -7,7 +7,7 @@ const defaults = {
   links: true,
   paragraphs: true
 }
-const toWiki = function(options) {
+const toWiki = function (options) {
   options = options || {}
   options = Object.assign({}, defaults, options)
   let text = ''
@@ -20,28 +20,21 @@ const toWiki = function(options) {
   //render infoboxes (up at the top)
   if (options.infoboxes === true) {
     text += this.infoboxes()
-      .map(i => i.wikitext(options))
+      .map((i) => i.wikitext(options))
       .join('\n')
   }
 
   //render each section
   if (options.sections === true || options.paragraphs === true || options.sentences === true) {
     let sections = this.sections()
-    sections = sections.filter(s => s.title() !== 'References')
-    text += sections.map(s => s.wikitext(options)).join('\n')
-  }
-  //default off
-  if (options.references === true) {
-    text += '== References =='
-    text += this.references()
-      .map(c => c.wiki(options))
-      .join('\n')
+    // sections = sections.filter((s) => s.title() !== 'References')
+    text += sections.map((s) => s.wikitext(options)).join('\n')
   }
 
   // add categories on the bottom
   if (options.categories === true) {
     text += '\n'
-    this.categories().forEach(cat => (text += `\n[[${cat}]]`))
+    this.categories().forEach((cat) => (text += `\n[[Category: ${cat}]]`))
   }
   return text
 }