+++ plugins/wikitext/tests/misc.test.js
@@ -1,16 +1,21 @@
 var test = require('tape')
 var wtf = require('./_lib')
 
-var tidy = str => {
+var tidy = (str) => {
   str = str.replace(/\s[\s]+/g, ' ')
   str = str.replace(/\n/g, '')
   str = str.trim()
   return str
 }
 
-test('basic-wikitext', t => {
-  let arr = ['that cat is [[a]] cool dude', `one [[two|2]] three`, 'cool [[stuff]] **bold** too']
-  arr.forEach(str => {
+test('basic-wikitext', (t) => {
+  let arr = [
+    'that cat is [[a]] cool dude',
+    `one [[two|2]] three`,
+    'cool [[stuff]] **bold** too',
+    '{{foobar | fun = true | key = val}}'
+  ]
+  arr.forEach((str) => {
     let doc = wtf(str)
     t.equal(tidy(doc.wikitext()), tidy(str), str)
   })